Fix compiler warnings and clean up

+++ sam/src/SckBase.h
@@ -159,7 +159,7 @@ class SckBase
 		uint8_t wakeUP_M = 0;
 		void updatePower();
 		uint32_t updatePowerMillis = 0;
-		void goToSleep(uint16_t sleepPeriod=3000); 	// sleepPeriod in ms
+		void goToSleep(uint32_t sleepPeriod=3000); 	// sleepPeriod in ms
 		void configGCLK6(); 			// Taken from https://github.com/arduino-libraries/ArduinoLowPower
 
 		// **** Sensors
